+++ testing/lib/sanityChecks.js
